+++ acvillager/src/components/ui/Header.js
@@ -1,11 +1,10 @@
 import React from 'react'
-import headerimg from '../../../images/headerimg.png';
-import './Header.css'
+import headerimg from '../../images/headerimg.png';
 
 
 function AppHeader() {
     return (
-        <header className="App-header">
+        <header className="app-header">
             <img src={headerimg} alt='Animal Crossing logo'></img>
         </header>
     )
