--- acvillager/src/components/ui/Header/Header.css
@@ -1,4 +1,0 @@
-.App-header img {
-  width: 20%;
-  height: auto;
-}
